Split set price process in three parts to minimize reverting efforts when needed, separate product prices from product keep changes at a point at a time, change int to long as product quantity unit, add multiple instances for test data

+++ calculator/src/main/java/com/market/banica/calculator/controller/CalculatorController.java
@@ -32,7 +32,7 @@ public class CalculatorController {
     @GetMapping("/{clientId}/{itemName}/{quantity}")
     public List<ProductDto> getRecipe(@PathVariable("clientId") @NotBlank String clientId,
                                                 @PathVariable("itemName") @NotBlank String itemName,
-                                                @PathVariable("quantity") @Min(1) int quantity) {
+                                                @PathVariable("quantity") @Min(1) long quantity) {
         LOGGER.info("GET /calculator called");
